Simplify entity extraction
