--- ts/packages/agents/browser/src/extension/views/entityComparison.ts
@@ -1,187 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-// Entity Comparison Implementation
-// Advanced entity comparison and analysis features
-
-export interface EntityComparison {
-    entities: any[];
-    commonRelationships: any[];
-    uniqueRelationships: { [entityName: string]: any[] };
-    similarityScore: number;
-    comparisonMetrics: EntityComparisonMetrics;
-    sharedDomains: string[];
-    sharedTopics: string[];
-    temporalOverlap: TemporalOverlap;
-}
-
-export interface EntityComparisonMetrics {
-    totalRelationships: { [entityName: string]: number };
-    averageConfidence: { [entityName: string]: number };
-    averageStrength: { [entityName: string]: number };
-    centralityScore: { [entityName: string]: number };
-    influenceScore: { [entityName: string]: number };
-}
-
-export interface TemporalOverlap {
-    overlapPeriod: string;
-    overlapPercentage: number;
-    commonTimeframe: {
-        start: string;
-        end: string;
-    };
-}
-
-export interface EntityClusterAnalysis {
-    clusters: EntityCluster[];
-    entityClusterMembership: { [entityName: string]: string };
-    crossClusterConnections: number;
-    clusterCoherence: { [clusterId: string]: number };
-}
-
-export interface EntityCluster {
-    clusterId: string;
-    entities: string[];
-    centerEntity: string;
-    clusterType: "functional" | "temporal" | "topical" | "organizational";
-    coherenceScore: number;
-    description: string;
-}
-
-/**
- * Entity Comparison Component
- * Handles comparing multiple entities and analyzing their relationships
- */
-export class EntityComparisonManager {
-    private comparisonPanel: HTMLElement | null = null;
-    private selectedEntities: any[] = [];
-    private currentComparison: EntityComparison | null = null;
-    private maxComparisonEntities: number = 5;
-
-    constructor() {
-        this.initializeComparisonPanel();
-        this.setupComparisonEventHandlers();
-    }
-
-    /**
-     * Initialize comparison panel
-     */
-    private initializeComparisonPanel(): void {
-        this.createComparisonPanel();
-        this.setupComparisonControls();
-    }
-
-    /**
-     * Show comparison panel
-     */
-    showComparisonPanel(): void {
-        if (this.comparisonPanel) {
-            this.comparisonPanel.style.display = "block";
-        }
-    }
-
-    /**
-     * Hide comparison panel
-     */
-    hideComparisonPanel(): void {
-        if (this.comparisonPanel) {
-            this.comparisonPanel.style.display = "none";
-        }
-    }
-
-    /**
-     * Start comparison with specific entities
-     */
-    startComparison(entities: any[]): void {
-        this.selectedEntities = entities.slice(0, this.maxComparisonEntities);
-        this.showComparisonPanel();
-
-        if (this.selectedEntities.length >= 2) {
-            this.performComparison();
-        }
-    }
-
-    /**
-     * Perform entity comparison (simplified for now)
-     */
-    private async performComparison(): Promise<void> {
-        console.log("Performing comparison for:", this.selectedEntities);
-        // Implementation would go here
-    }
-
-    /**
-     * Create comparison panel (simplified)
-     */
-    private createComparisonPanel(): void {
-        this.comparisonPanel = document.createElement("div");
-        this.comparisonPanel.id = "entityComparisonPanel";
-        this.comparisonPanel.className = "entity-comparison-panel";
-        this.comparisonPanel.innerHTML = `
-            <div class="panel-header">
-                <h3>Entity Comparison</h3>
-                <button class="close-btn" id="closeComparisonBtn">
-                    <i class="bi bi-x"></i>
-                </button>
-            </div>
-            <div class="panel-content">
-                <p>Entity comparison functionality coming soon...</p>
-            </div>
-        `;
-
-        this.comparisonPanel.style.cssText = `
-            position: fixed;
-            left: 20px;
-            top: 50%;
-            transform: translateY(-50%);
-            width: 400px;
-            max-height: 80vh;
-            overflow-y: auto;
-            background: white;
-            border: 1px solid #ddd;
-            border-radius: 8px;
-            box-shadow: 0 4px 12px rgba(0,0,0,0.15);
-            z-index: 1002;
-            display: none;
-            padding: 16px;
-        `;
-
-        document.body.appendChild(this.comparisonPanel);
-    }
-
-    /**
-     * Set up comparison controls
-     */
-    private setupComparisonControls(): void {
-        const closeBtn = document.getElementById("closeComparisonBtn");
-        if (closeBtn) {
-            closeBtn.addEventListener("click", () =>
-                this.hideComparisonPanel(),
-            );
-        }
-    }
-
-    /**
-     * Set up comparison event handlers
-     */
-    private setupComparisonEventHandlers(): void {
-        // Listen for comparison requests
-        document.addEventListener("requestEntityComparison", (e: any) => {
-            const entities = e.detail.entities;
-            this.startComparison(entities);
-        });
-    }
-
-    /**
-     * Get selected entities
-     */
-    getSelectedEntities(): any[] {
-        return [...this.selectedEntities];
-    }
-
-    /**
-     * Get current comparison
-     */
-    getCurrentComparison(): EntityComparison | null {
-        return this.currentComparison;
-    }
-}
