+++ ts/packages/agents/browser/src/extension/views/entitySidebar.ts
@@ -303,142 +303,10 @@ export class EntitySidebar {
         }
     }
 
-    private generateMockAliases(name: string, type: string): string[] {
-        const aliases: string[] = [];
 
-        if (type === "technology") {
-            if (name === "React") {
-                aliases.push("React.js", "ReactJS");
-            } else if (name === "Next.js") {
-                aliases.push("NextJS", "Next");
-            } else if (name === "TypeScript") {
-                aliases.push("TS");
-            }
-        } else if (type === "organization") {
-            if (name === "Facebook") {
-                aliases.push("Meta", "Meta Platforms");
-            } else if (name === "Apple") {
-                aliases.push("Apple Inc.", "AAPL");
-            }
-        } else if (type === "person") {
-            if (name === "Satya Nadella") {
-                aliases.push("@satyanadella", "Satya Narayana Nadella");
-            }
-        }
 
-        return aliases;
-    }
 
-    private generateMockDomains(type: string): string[] {
-        const domainMap: { [key: string]: string[] } = {
-            technology: [
-                "github.com",
-                "stackoverflow.com",
-                "docs.microsoft.com",
-            ],
-            organization: [
-                "linkedin.com",
-                "techcrunch.com",
-                "bloomberg.com",
-                "cnbc.com",
-            ],
-            person: ["linkedin.com", "wikipedia.org"],
-            product: ["productHunt.com", "techreview.com", "wired.com"],
-        };
 
-        return domainMap[type] || ["example.com"];
-    }
-
-    private generateMockTopics(type: string): string[] {
-        const topicMap: { [key: string]: string[] } = {
-            technology: [
-                "software development",
-                "web development",
-                "programming",
-            ],
-            organization: ["business", "technology", "innovation"],
-            person: ["leadership", "entrepreneurship", "technology"],
-            product: ["consumer electronics", "technology", "design"],
-        };
-
-        return topicMap[type] || ["general"];
-    }
-
-    private generateMockDate(daysAgo: number): string {
-        const date = new Date();
-        date.setDate(date.getDate() + daysAgo);
-        return date.toISOString();
-    }
-
-    private generateMockRelationships(
-        entityName: string,
-        entityType: string,
-    ): any[] {
-        // Generate contextual mock relationships based on entity
-        const relationships: any[] = [];
-
-        if (entityName === "React") {
-            relationships.push(
-                {
-                    relatedEntity: "Facebook",
-                    relationshipType: "created_by",
-                    strength: 0.9,
-                    confidence: 0.95,
-                },
-                {
-                    relatedEntity: "Next.js",
-                    relationshipType: "framework_for",
-                    strength: 0.8,
-                    confidence: 0.9,
-                },
-                {
-                    relatedEntity: "TypeScript",
-                    relationshipType: "commonly_used_with",
-                    strength: 0.7,
-                    confidence: 0.85,
-                },
-            );
-        } else if (entityName === "Microsoft") {
-            relationships.push(
-                {
-                    relatedEntity: "Satya Nadella",
-                    relationshipType: "led_by",
-                    strength: 0.9,
-                    confidence: 0.95,
-                },
-                {
-                    relatedEntity: "Azure",
-                    relationshipType: "develops",
-                    strength: 0.95,
-                    confidence: 0.98,
-                },
-                {
-                    relatedEntity: "Office365",
-                    relationshipType: "develops",
-                    strength: 0.6,
-                    confidence: 0.8,
-                },
-            );
-        } else {
-            // Generate generic relationships
-            relationships.push(
-                {
-                    relatedEntity: "Related Entity 1",
-                    relationshipType: "related_to",
-                    strength: 0.6,
-                    confidence: 0.7,
-                },
-                {
-                    relatedEntity: "Related Entity 2",
-                    relationshipType: "connected_to",
-                    strength: 0.4,
-                    confidence: 0.6,
-                },
-            );
-        }
-
-        return relationships;
-    }
 
     private getEntityIcon(type: string): string {
         const iconMap: { [key: string]: string } = {
