--- ts/packages/agents/browser/src/extension/views/multiHopExplorer.ts
@@ -1,317 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-// Multi-hop Entity Exploration Implementation
-// Advanced entity network expansion and multi-hop exploration
-
-import {
-    EntityGraphServices,
-    DefaultEntityGraphServices,
-} from "./knowledgeUtilities";
-
-export interface EntityExpansionData {
-    centerEntity: string;
-    entities: any[];
-    relationships: any[];
-    depth: number;
-    expansionType: "breadth_first" | "depth_first" | "importance_based";
-}
-
-export interface ExpansionHistory {
-    step: number;
-    action: "expand" | "collapse" | "focus";
-    entity: string;
-    timestamp: number;
-    resultingEntityCount: number;
-}
-
-export interface NetworkMetrics {
-    totalNodes: number;
-    totalEdges: number;
-    density: number;
-    averageDegree: number;
-    maxDegree: number;
-    centralMostEntity: string;
-    clusters: number;
-}
-
-/**
- * Multi-hop Entity Exploration Component
- * Handles expanding entity networks, multi-hop traversal, and network analysis
- */
-export class MultiHopExplorer {
-    private visualizer: any; // EnhancedEntityGraphVisualizer reference
-    private currentDepth: number = 1;
-    private maxDepth: number = 3;
-    private expansionHistory: ExpansionHistory[] = [];
-    private networkCache: Map<string, EntityExpansionData> = new Map();
-    private isExpanding: boolean = false;
-    private entityGraphService: EntityGraphServices;
-
-    constructor(visualizer: any, entityGraphService?: EntityGraphServices) {
-        this.visualizer = visualizer;
-        this.entityGraphService =
-            entityGraphService || new DefaultEntityGraphServices();
-        this.setupExpansionControls();
-        this.setupExpansionEventHandlers();
-    }
-
-    /**
-     * Set up expansion control UI
-     */
-    private setupExpansionControls(): void {
-        this.setupExpansionButtons();
-        this.setupDepthControls();
-        this.setupExpansionSettings();
-        this.setupHistoryControls();
-    }
-
-    /**
-     * Set up expansion buttons
-     */
-    private setupExpansionButtons(): void {
-        const expandBtn = document.getElementById("expandBtn");
-        const collapseBtn = document.getElementById("collapseBtn");
-        const resetBtn = document.getElementById("resetBtn");
-
-        if (expandBtn) {
-            expandBtn.addEventListener("click", () =>
-                this.expandSelectedEntities(),
-            );
-        }
-
-        if (collapseBtn) {
-            collapseBtn.addEventListener("click", () =>
-                this.collapseSelectedEntities(),
-            );
-        }
-
-        if (resetBtn) {
-            resetBtn.addEventListener("click", () => this.resetGraph());
-        }
-    }
-
-    /**
-     * Expand entity network
-     */
-    async expandEntityNetwork(
-        entityName: string,
-        depth: number = 1,
-    ): Promise<void> {
-        try {
-            // Check cache first
-            const cacheKey = `${entityName}-${depth}`;
-            let expansionData = this.networkCache.get(cacheKey);
-
-            if (!expansionData) {
-                // Generate expansion data
-                expansionData = await this.generateExpansionData(
-                    entityName,
-                    depth,
-                );
-                this.networkCache.set(cacheKey, expansionData);
-            }
-
-            // Add new entities and relationships to visualizer
-            await this.addEntitiesAndRelationships(expansionData);
-
-            // Apply layout if auto-layout is enabled
-            const autoLayout = document.getElementById(
-                "autoLayout",
-            ) as HTMLInputElement;
-            if (autoLayout?.checked) {
-                this.visualizer.applyCurrentLayout();
-            }
-
-            // Update metrics
-            this.updateNetworkMetrics();
-        } catch (error) {
-            console.error("Failed to expand entity network:", error);
-            throw error;
-        }
-    }
-
-    /**
-     * Expand selected entities
-     */
-    async expandSelectedEntities(): Promise<void> {
-        if (this.isExpanding) return;
-
-        const selectedEntities = this.visualizer.getSelectedEntities();
-        if (selectedEntities.length === 0) {
-            this.showMessage("Please select entities to expand", "warning");
-            return;
-        }
-
-        this.isExpanding = true;
-        this.showExpansionProgress(
-            `Expanding ${selectedEntities.length} entities...`,
-        );
-
-        try {
-            for (const entityName of selectedEntities) {
-                await this.expandEntityNetwork(entityName, this.currentDepth);
-            }
-
-            this.addToHistory("expand", selectedEntities.join(", "));
-            this.showMessage(
-                `Expanded ${selectedEntities.length} entities successfully`,
-                "success",
-            );
-        } catch (error) {
-            console.error("Expansion failed:", error);
-            this.showMessage("Expansion failed. Please try again.", "error");
-        } finally {
-            this.isExpanding = false;
-            this.hideExpansionProgress();
-        }
-    }
-
-    // Additional simplified methods for basic functionality
-    private setupDepthControls(): void {
-        console.log("Setting up depth controls");
-    }
-
-    private setupExpansionSettings(): void {
-        console.log("Setting up expansion settings");
-    }
-
-    private setupHistoryControls(): void {
-        console.log("Setting up history controls");
-    }
-
-    private setupExpansionEventHandlers(): void {
-        console.log("Setting up expansion event handlers");
-    }
-
-    private async generateExpansionData(
-        entityName: string,
-        depth: number,
-    ): Promise<EntityExpansionData> {
-        return this.generateRealExpansionData(entityName, depth);
-    }
-
-    /**
-     * Generate real expansion data using enhanced search
-     */
-    private async generateRealExpansionData(
-        entityName: string,
-        depth: number,
-    ): Promise<EntityExpansionData> {
-        try {
-            // Get entity graph from enhanced search
-            const graphData = await this.entityGraphService.getEntityGraph(
-                entityName,
-                depth,
-            );
-
-            // Convert to expansion data format
-            return {
-                centerEntity: entityName,
-                entities: graphData.entities.map((e: any) => ({
-                    name: e.name,
-                    type: e.type,
-                    confidence: e.confidence,
-                })),
-                relationships: graphData.relationships.map((r: any) => ({
-                    from: r.relatedEntity,
-                    to: entityName, // Assuming relationships point to center entity
-                    type: r.relationshipType,
-                    strength: r.strength,
-                })),
-                depth: depth,
-                expansionType: "importance_based",
-            };
-        } catch (error) {
-            console.error("Failed to generate real expansion data:", error);
-            // Fallback to mock data
-            return this.generateMockExpansionData(entityName, depth);
-        }
-    }
-
-    /**
-     * Generate mock expansion data
-     */
-    private generateMockExpansionData(
-        entityName: string,
-        depth: number,
-    ): EntityExpansionData {
-        // Mock data generation
-        return {
-            centerEntity: entityName,
-            entities: [
-                {
-                    name: `${entityName} Related 1`,
-                    type: "organization",
-                    confidence: 0.8,
-                },
-                {
-                    name: `${entityName} Related 2`,
-                    type: "person",
-                    confidence: 0.7,
-                },
-            ],
-            relationships: [
-                {
-                    from: entityName,
-                    to: `${entityName} Related 1`,
-                    type: "related_to",
-                    strength: 0.8,
-                },
-                {
-                    from: entityName,
-                    to: `${entityName} Related 2`,
-                    type: "connected_to",
-                    strength: 0.7,
-                },
-            ],
-            depth: depth,
-            expansionType: "importance_based",
-        };
-    }
-
-    private async addEntitiesAndRelationships(
-        expansionData: EntityExpansionData,
-    ): Promise<void> {
-        if (this.visualizer && this.visualizer.addElements) {
-            await this.visualizer.addElements({
-                entities: expansionData.entities,
-                relationships: expansionData.relationships,
-            });
-        }
-    }
-
-    private async collapseSelectedEntities(): Promise<void> {
-        console.log("Collapsing selected entities");
-    }
-
-    private async resetGraph(): Promise<void> {
-        console.log("Resetting graph");
-    }
-
-    private updateNetworkMetrics(): void {
-        console.log("Updating network metrics");
-    }
-
-    private addToHistory(
-        action: "expand" | "collapse" | "focus",
-        entity: string,
-    ): void {
-        console.log(`Adding to history: ${action} ${entity}`);
-    }
-
-    private showExpansionProgress(message: string): void {
-        console.log("Expansion progress:", message);
-    }
-
-    private hideExpansionProgress(): void {
-        console.log("Expansion complete");
-    }
-
-    private showMessage(
-        message: string,
-        type: "success" | "warning" | "error",
-    ): void {
-        console.log(`${type.toUpperCase()}: ${message}`);
-    }
-}
